backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv
requests
beautifulsoup4
celery
redis
sqlalchemy
openai
openai-agents
google-search-results
faiss-cpu
sentence-transformers
scikit-learn
torch
faiss-cpu
sentence-transformers
scikit-learn
torch
